Move _document_directive to directives.py.
